fix: refactor api spec runner into some helpers and rule in its own file
